+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- added autodetection for MX3000D inverters (SNR: 'Y00')
+
 ### Changed
 
 - Update dependency coverage to v7.14.0