+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- speed up unit tests
 - improve MODBUS scan with a length check of the responses
 - Update dependency coverage to v7.14.0
 - Update dependency python to v3.14.5