+++ CHANGELOG.md
@@ -7,8 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- sensor_list detection for client mode
+- push MQTT inverter values on starting the client mode
+- added INSULATION_IMP_RX and RY values for MX3000D inverters
+- added autodetection for MX3000D inverters (SNR: 'Y00')
+
 ### Changed
 
+- speed up unit tests
+- improve MODBUS scan with a length check of the responses
 - Update dependency coverage to v7.14.0
 - Update dependency python to v3.14.5